+++ src/store/reducers/local-data/local-data.test.js
@@ -1,75 +1,18 @@
-import MockAdapter from "axios-mock-adapter";
-import createAPI from "../api";
-import {
-  reducer,
-  isGenreAnswerCorrect,
-  ActionCreator,
-  Operation
-} from "./local-data";
 
-describe(`Business logic is correct`, () => {
-  it(`Genre question is  checked correctly`, () => {
-    expect(isGenreAnswerCorrect([false, true, false, false], {
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    })).toEqual(true);
+import localData from "./local-data";
 
-    expect(isGenreAnswerCorrect([false, false, true, false], {
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    })).toEqual(false);
-  });
-
-});
 
 describe(`Reducer works correctly`, () => {
   it(`Reducer without additional parameters should return initial state`, () => {
-    expect(reducer(undefined, {})).toEqual({
+    expect(localData(undefined, {})).toEqual({
       step: -1,
-      questions: [],
       mistakes: 0,
       time: 300000,
       gameTimer: null
     });
   });
   it(`Reducer should increment current step by a given value`, () => {
-    expect(reducer({
+    expect(localData({
       step: -1,
       mistakes: 0,
     }, {
@@ -80,7 +23,7 @@ describe(`Reducer works correctly`, () => {
       mistakes: 0,
     });
 
-    expect(reducer({
+    expect(localData({
       step: -1,
       mistakes: 0,
     }, {
@@ -93,7 +36,7 @@ describe(`Reducer works correctly`, () => {
   });
 
   it(`Reducer should increment number of mistakes by a given value`, () => {
-    expect(reducer({
+    expect(localData({
       step: -1,
       mistakes: 0,
     }, {
@@ -104,7 +47,7 @@ describe(`Reducer works correctly`, () => {
       mistakes: 1,
     });
 
-    expect(reducer({
+    expect(localData({
       step: -1,
       mistakes: 0,
     }, {
@@ -117,9 +60,8 @@ describe(`Reducer works correctly`, () => {
   });
 
   it(`Reducer should correctly reset application state`, () => {
-    expect(reducer({
+    expect(localData({
       step: 898,
-      questions: [],
       mistakes: 34,
       time: 300000,
       gameTimer: null
@@ -128,7 +70,6 @@ describe(`Reducer works correctly`, () => {
       type: `RESET`,
     })).toEqual({
       step: -1,
-      questions: [],
       mistakes: 0,
       time: 300000,
       gameTimer: null
@@ -136,7 +77,7 @@ describe(`Reducer works correctly`, () => {
   });
 
   it(`Reducer correctly decrements time`, () => {
-    expect(reducer({
+    expect(localData({
       questionStep: -1,
       mistakes: 0,
       time: 300000,
@@ -155,315 +96,4 @@ describe(`Reducer works correctly`, () => {
 
 });
 
-describe(`load data test group`, () => {
-  it(`Should make a correct call to /questions`, () => {
-    const dispatch = jest.fn();
-    const api = createAPI(dispatch);
-    const apiMock = new MockAdapter(api);
-    const questionLoader = Operation.loadQuestions();
 
-    apiMock
-      .onGet(`/questions`)
-      .reply(200, [{lul: true}]);
-
-    return questionLoader(dispatch, null, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: `LOAD_QUESTIONS`,
-          payload: [{lul: true}]
-        });
-      });
-  });
-
-
-});
-
-describe(`Action creators work correctly`, () => {
-  it(`Action creator for incrementing step return correct action`, () => {
-    expect(ActionCreator.incrementStep()).toEqual({
-      type: `INCREMENT_STEP`,
-      payload: 1,
-    });
-  });
-
-  it(`Action creator for incrementing mistakes return action with 0 payload`, () => {
-    expect(ActionCreator.incrementMistake({
-      artist: `correct`,
-      picture: ``
-    }, [{
-      type: `artist`,
-      song: {
-        artist: `correct`,
-        src: ``,
-      }, answers: [{
-        artist: `correct`,
-        picture: ``
-      }, {
-        artist: `incorrect`,
-        picture: ``
-      }, {
-        artist: `incorrect-2`,
-        picture: ``
-      }]
-    }, {
-      type: `artist`,
-      song: {
-        artist: `correct`,
-        src: ``,
-      }, answers: [{
-        artist: `correct`,
-        picture: ``
-      }, {
-        artist: `incorrect`,
-        picture: ``
-      }, {
-        artist: `incorrect-2`,
-        picture: ``
-      }]
-    }], 0, Infinity, 0)).toEqual({
-      type: `INCREMENT_MISTAKES`,
-      payload: 0,
-    });
-  });
-
-  it(`Action creator for incrementing mistakes return action with 1 payload`, () => {
-    expect(ActionCreator.incrementMistake({
-      artist: `incorrect`,
-      picture: ``
-    }, [{
-      type: `artist`,
-      song: {
-        artist: `correct`,
-        src: ``,
-      }, answers: [{
-        artist: `correct`,
-        picture: ``
-      }, {
-        artist: `incorrect`,
-        picture: ``
-      }, {
-        artist: `incorrect-2`,
-        picture: ``
-      }]
-    }, {
-      type: `artist`,
-      song: {
-        artist: `correct`,
-        src: ``,
-      }, answers: [{
-        artist: `correct`,
-        picture: ``
-      }, {
-        artist: `incorrect`,
-        picture: ``
-      }, {
-        artist: `incorrect-2`,
-        picture: ``
-      }]
-    }], 0, Infinity, 0)).toEqual({
-      type: `INCREMENT_MISTAKES`,
-      payload: 1,
-    });
-  });
-
-  it(`Action creator for incrementing mistakes return action with 0 payload`, () => {
-    expect(ActionCreator.incrementMistake([false, true, false, false], [{
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    },
-    {
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    }], 0, Infinity, 0)).toEqual({
-      type: `INCREMENT_MISTAKES`,
-      payload: 0,
-    });
-  });
-
-  it(`Action creator for incrementing mistakes return action with 1 payload`, () => {
-    expect(ActionCreator.incrementMistake([false, false, true, false], [{
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    },
-    {
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    }], 0, Infinity, 0)).toEqual({
-      type: `INCREMENT_MISTAKES`,
-      payload: 1,
-    });
-  });
-
-  it(`Action creator reset state if user is answer incorrectly and sum mistakes = max`, () => {
-    expect(ActionCreator.incrementMistake({
-      artist: `incorrect`,
-      picture: ``
-    }, [{
-      type: `artist`,
-      song: {
-        artist: `correct`,
-        src: ``,
-      }, answers: [{
-        artist: `correct`,
-        picture: ``
-      }, {
-        artist: `incorrect`,
-        picture: ``
-      }, {
-        artist: `incorrect-2`,
-        picture: ``
-      }]
-    }], Infinity, 0, 0)).toEqual({
-      type: `RESET`,
-    });
-  });
-
-  it(`Action creator reset state if user is answer incorrectly and sum mistakes = max`, () => {
-    expect(ActionCreator.incrementMistake([false, false, true, false], [{
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    }], Infinity, 0, 0)).toEqual({
-      type: `RESET`
-    });
-  });
-
-  it(`Action creator reset state if step more then lenght array questions`, () => {
-    expect(ActionCreator.incrementMistake([false, true, false, false], [{
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    }, {
-      type: `genre`,
-      genre: `dance`,
-      answers: [
-        {
-          src: `1`,
-          genre: `rock`,
-        },
-        {
-          src: `2`,
-          genre: `dance`,
-        },
-        {
-          src: `3`,
-          genre: `jazz`,
-        },
-        {
-          src: `4`,
-          genre: `rock`,
-        },
-      ],
-    }], Infinity, 0, 3)).toEqual({
-      type: `RESET`
-    });
-  });
-
-});
